requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
idna==3.4
lxml==4.9.3
markdown-it-py==3.0.0
mdurl==0.1.2
mysql==0.0.3
mysqlclient==2.2.0
Pillow==10.0.1
Pygments==2.16.1
pytz==2023.3.post1
requests==2.31.0
rich==13.5.3
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.5
